requirements.txt
requests
python-dotenv
customtkinter
Pillow